requirements.txt
numpy==2.2.6
faiss-cpu==1.12.0
sentence-transformers==5.1.1
PyPDF2==3.0.1
langchain==0.3.27
pickle-mixin==1.0.2
streamlit==1.50.0
ollama==0.6.0
torch==2.9.0
transformers==4.57.1
scikit-learn==1.7.2
tqdm==4.67.1
requests==2.32.5
pandas==2.3.3
